requirements.txt
asyncpg==0.30.0
python-dotenv==1.1.1